twisted/requirements.txt
Twisted==15.5.0
argparse==1.2.1
cffi
characteristic==14.3.0
cryptography
enum34==1.1.2
idna==2.0
ipaddress==1.0.16
pika==0.10.0
pyOpenSSL
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
requests==2.9.1
service-identity==14.0.0
six==1.10.0
treq==15.1.0
wsgiref==0.1.2
zope.interface==4.1.3
PyYAML
